[table] fix cycles
